requirements.txt
# Israeli Basketball Calendar - Python Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP Client
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0

# Caching
cachetools>=5.3.0

# Date/Time
python-dateutil>=2.8.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx  # for TestClient

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
